Update triggering events for GitHub workflows
This commit modifies the event triggers in maven-publish.yml and quality-gate.yml workflows. The triggering event has been changed from 'repository_dispatch' to 'push' events on 'main' and 'develop' branches and the code for triggering another workflow has been removed. This will ensure that the workflows are correctly triggered when changes are made to these branches.